refactor(mc): split terminal.js layout half into TerminalLayoutBase (#878)
Clear the S104 (>500 lines) smell on terminal.js (862 -> 522 lines) by moving
the tab / split-pane layout half of TerminalManager into a TerminalLayoutBase
class in terminal-layout.js; TerminalManager now `extends` it, so every method
resolves through the prototype chain with no call-site changes. Both stay plain
(non-module) scripts: terminal-layout.js loads first and publishes the base
class + shared status constants on globalThis, which the browser script tags and
the jest `require` harness both resolve. eslint clean; 227 jest tests pass.
